requirements.txt
Appium-Python-Client
pytest
selenium
pytest-rerunfailures
behave
allure-behave
pytest-xdist
request